pyproject.toml
[tool.poetry]
name = "django-auth"
version = "0.1.0"
description = ""
authors = ["Prajwal k <prajwalk@unicourt.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
djangorestframework = "^3.13.1"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
